python-app/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-distro[otlp]
opentelemetry-exporter-otlp
